+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Migrations trigger during module activation
 - Invoice is now downloaded with dynamic filename (invoice number is included)
+- Donwloaded file name prefix ins configurable in settings
 - Currency shown near prices in the invoice
 - Footer for invoice document is modifiable in module settings